Fix generative design overflows

+++ generative_design/shape/p_2_2_1_02.rs
@@ -54,7 +54,8 @@ struct Model {
     radius: f32,
     direction: Direction,
     draw_mode: u8,
-    counter: u8,
+    // p5js counter max ~9 quadrillion... our max ~18 quintillion
+    counter: u64,
 }
 
 fn model(app: &App) -> Model {
